# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EdTech Study Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chapters/{chapterId}/topics:
    get:
      operationId: getTopics
      tags: [topics]
      summary: List topics for a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
    post:
      operationId: createTopic
      tags: [topics]
      summary: Create a topic (admin only)
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
  /topics/{topicId}:
    patch:
      operationId: updateTopic
      tags: [topics]
      summary: Update a topic (admin only)
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"

  # ── Progress (Gate Checker) ───────────────────────────────────────────
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get today's study tasks
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyTask"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a manual study task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyTask"
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task status or details
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyTaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyTask"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── External Tests ────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Topic:
      type: object
      properties:
        id:
          type: string
        chapter_id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        order_index:
          type: integer
        is_active:
          type: boolean
        telegram_chat_id:
          type: ["string", "null"]
        telegram_message_id:
          type: ["string", "null"]
      required: [id, chapter_id, title, order_index, is_active]
    TopicInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        order_index:
          type: integer
        telegram_chat_id:
          type: ["string", "null"]
        telegram_message_id:
          type: ["string", "null"]
      required: [title, order_index]
    TopicUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        order_index:
          type: integer
        telegram_chat_id:
          type: ["string", "null"]
        telegram_message_id:
          type: ["string", "null"]
        is_active:
          type: boolean

    # ── Gate / Progress ───────────────────────────────────────────────────
...
    StudyTask:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        target_type:
          type: string
          enum: [platform_subtopic, personal_topic, free_text]
        target_id:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        priority:
          type: integer
        due_date:
          type: ["string", "null"]
        source:
          type: string
          enum: [auto, manual]
        created_at:
          type: string
      required: [id, title, target_type, status, priority, source, created_at]
    StudyTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        target_type:
          type: string
          enum: [platform_subtopic, personal_topic, free_text]
        target_id:
          type: ["string", "null"]
        priority:
          type: integer
        due_date:
          type: ["string", "null"]
      required: [title, target_type]
    StudyTaskUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        priority:
          type: integer

    # ── External Tests ────────────────────────────────────────────────────
...