# lib/api-zod/src/generated/types/studyTask.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EdTech Study Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StudyTaskSource } from './studyTaskSource';
import type { StudyTaskStatus } from './studyTaskStatus';
import type { StudyTaskTargetType } from './studyTaskTargetType';

export interface StudyTask {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  target_type: StudyTaskTargetType;
  /** @nullable */
  target_id?: string | null;
  status: StudyTaskStatus;
  priority: number;
  /** @nullable */
  due_date?: string | null;
  source: StudyTaskSource;
  created_at: string;
}