# lib/api-zod/src/generated/types/studyTaskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EdTech Study Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StudyTaskInputTargetType } from './studyTaskInputTargetType';

export interface StudyTaskInput {
  title: string;
  /** @nullable */
  description?: string | null;
  target_type: StudyTaskInputTargetType;
  /** @nullable */
  target_id?: string | null;
  priority?: number;
  /** @nullable */
  due_date?: string | null;
}